Create build_hpo.py to generate hpo network and avoid circular ref
