+++ tests/test_util.py
@@ -4,7 +4,9 @@ import unittest
 from phenopy.util import parse, read_records_file
 
 from phenopy.config import config
-from phenopy.util import read_phenotype_groups, encode_phenotypes, parse_input, generate_annotated_hpo_network
+from phenopy.util import read_phenotype_groups, encode_phenotypes, parse_input
+from phenopy.build_hpo import generate_annotated_hpo_network
+
 
 
 class UtilTestCase(unittest.TestCase):
